+++ packages/deptree/test/fixtures/file/basic/sub2.ts
@@ -1,0 +1,3 @@
+export const id = 'sub2'
+
+export const sub2Func = () => {}