+++ packages/deptree/test/fixtures/module/circularDeps/sub1.ts
@@ -1,0 +1,1 @@
+export function sub1() {}