+++ packages/deptree/test/fixtures/module/ignoreUnresolvable/input.ts
@@ -1,0 +1,4 @@
+import { sub } from './sub'
+import unresolvable from './unresolvable'
+
+sub()