+++ packages/deptree/test/fixtures/module/jsxWithDecorator/Sub2/helper.ts
@@ -1,0 +1,1 @@
+export const calculate = (x: number) => Math.sqrt(x)