+++ packages/deptree/test/fixtures/module/mixingImportAndRequire/input.ts
@@ -1,0 +1,7 @@
+import { subFunc } from './sub1'
+
+const { id, sub2Func } = require('./sub2')
+
+console.log(id)
+subFunc()
+sub2Func()