+++ CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/vincenteof/reddeps/compare/v0.4.0...v0.4.1) (2022-08-08)
+
+**Note:** Version bump only for package reddeps
+
+# [0.4.0](https://github.com/vincenteof/reddeps/compare/v0.3.3...v0.4.0) (2022-08-08)
+
+### Features
+
+* ignore unresolvable ([e46e4a9](https://github.com/vincenteof/reddeps/commit/e46e4a9a7aea27ca44e81b4f5f97c7e864a9ef55))
+
+## [0.3.3](https://github.com/vincenteof/reddeps/compare/v0.3.2...v0.3.3) (2022-08-08)
+
+### Bug Fixes
+
+* resolve search dir ([68052a6](https://github.com/vincenteof/reddeps/commit/68052a670ebf18d7d584c90f75bf1a678b41be1d))
+
+## [0.3.2](https://github.com/vincenteof/reddeps/compare/v0.3.1...v0.3.2) (2022-08-08)
+
+**Note:** Version bump only for package reddeps
+
+## [0.3.1](https://github.com/vincenteof/reddeps/compare/v0.3.0...v0.3.1) (2022-08-08)
+
+**Note:** Version bump only for package reddeps
+
+# [0.3.0](https://github.com/vincenteof/reddeps/compare/v0.2.0...v0.3.0) (2022-08-08)
+
+### Features
+
+* basic deptree implementation ([7d19e0c](https://github.com/vincenteof/reddeps/commit/7d19e0c379aa906c4a45480526bca3a755d892c5))
+* cli with analyze and clean ([74f1a69](https://github.com/vincenteof/reddeps/commit/74f1a69938b35022b10c9956e9ca3d51c3f14007))
+* compatible with require ([3096267](https://github.com/vincenteof/reddeps/commit/30962679b9e5a7e6f416ca6280f35bc499a018d0))
+* findUnused api ([08fc295](https://github.com/vincenteof/reddeps/commit/08fc2955cb99e9bc1317634249600ebee066cd6a))
+* ignore patterns ([01f57e9](https://github.com/vincenteof/reddeps/commit/01f57e99297cc799c389214650798b1abfbe54d2))
+
 # [0.2.0](https://github.com/vincenteof/reddeps/compare/v0.1.0...v0.2.0) (2022-07-31)
 
 ### Features